requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.15
botocore==1.29.15
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-storages==1.13.1
djangorestframework==3.14.0
jmespath==1.0.1
kombu==5.2.4
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.5
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
